+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.16.5
+
+* Added `topdown=False` support to `walk()`
+* Added `copytree()` example
+
 ## 0.16.4
 
 * Creating a write-mode `BlobFile` for a local path will automatically create intermediate directories to be consistent with blob storage, see https://github.com/christopher-hesse/blobfile/issues/48